+++ src/components/articles/ArticleCardList/ArticleCardList.stories.tsx
@@ -1,23 +1,18 @@
-
-import { ArticleCardList } from './ArticleCardList'
+import { ArticleCardListPresenter } from './ArticleCardListPresenter'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
 const meta = {
-    component: ArticleCardList,
+    component: ArticleCardListPresenter,
     tags: ['autodocs'],
+    parameters: {},
     args: {
-        page: 0,
+        cardList: [],
+        totalPage: 1,
     },
-} satisfies Meta<typeof ArticleCardList>
+} satisfies Meta<typeof ArticleCardListPresenter>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
-export const Default: Story = {
-    render: (args) => (
-        <div className='max-w-[600px]'>
-            <ArticleCardList {...args} />
-        </div>
-    ),
-}
+export const Default: Story = {}
